+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 
 - Fixed occassional lag after loading the game, caused by Satsuma rigidbody being enabled when it should not be
 - Fixed "crashing" sounds coming from Satsuma and Jonnez when walking next to them
+- Fixed empty plastic cans possibly being unpickable by the player
+- Fixed MOP disabling empty items, even when "Disable Empty Items" checkbox was not ticked
 
 ## 3.3 (20.05.2021)